type

+++ ngx_pg_module.c
@@ -98,7 +98,7 @@ static ngx_int_t ngx_pg_parser_all(ngx_pg_save_t *s, const uintptr_t data) {
 }
 
 static ngx_int_t ngx_pg_parser_atttypmod(ngx_pg_save_t *s, const uintptr_t data) {
-    uint32_t atttypmod = (uint32_t)data;
+    unsigned long atttypmod = (unsigned long)data;
     ngx_log_debug1(NGX_LOG_DEBUG_HTTP, s->connection->log, 0, "%i", atttypmod);
     return NGX_OK;
 }
@@ -119,7 +119,7 @@ static ngx_int_t ngx_pg_parser_close(ngx_pg_save_t *s) {
 }
 
 static ngx_int_t ngx_pg_parser_columnid(ngx_pg_save_t *s, const uintptr_t data) {
-    uint16_t columnid = (uint16_t)data;
+    unsigned short columnid = (unsigned short)data;
     ngx_log_debug1(NGX_LOG_DEBUG_HTTP, s->connection->log, 0, "%i", columnid);
     return NGX_OK;
 }
@@ -140,7 +140,7 @@ static ngx_int_t ngx_pg_parser_tup(ngx_pg_save_t *s) {
 }
 
 static ngx_int_t ngx_pg_parser_nbytes(ngx_pg_save_t *s, const uintptr_t data) {
-    uint32_t nbytes = (uint32_t)data;
+    unsigned long nbytes = (unsigned long)data;
     ngx_log_debug1(NGX_LOG_DEBUG_HTTP, s->connection->log, 0, "%i", nbytes);
     return NGX_OK;
 }
@@ -167,7 +167,7 @@ static ngx_int_t ngx_pg_parser_name(ngx_pg_save_t *s, size_t len, const u_char *
 }
 
 static ngx_int_t ngx_pg_parser_format(ngx_pg_save_t *s, const uintptr_t data) {
-    uint16_t format = (uint16_t)data;
+    unsigned short format = (unsigned short)data;
     ngx_log_debug1(NGX_LOG_DEBUG_HTTP, s->connection->log, 0, "%i", format);
     return NGX_OK;
 }
@@ -188,19 +188,19 @@ static ngx_int_t ngx_pg_parser_intrans(ngx_pg_save_t *s) {
 }
 
 static ngx_int_t ngx_pg_parser_key(ngx_pg_save_t *s, const uintptr_t data) {
-    uint32_t key = (uint32_t)data;
+    unsigned long key = (unsigned long)data;
     ngx_log_debug1(NGX_LOG_DEBUG_HTTP, s->connection->log, 0, "%i", key);
     return NGX_OK;
 }
 
 static ngx_int_t ngx_pg_parser_method(ngx_pg_save_t *s, const uintptr_t data) {
-    uint32_t method = (uint32_t)data;
+    unsigned long method = (unsigned long)data;
     ngx_log_debug1(NGX_LOG_DEBUG_HTTP, s->connection->log, 0, "%i", method);
     return NGX_OK;
 }
 
 static ngx_int_t ngx_pg_parser_nfields(ngx_pg_save_t *s, const uintptr_t data) {
-    uint16_t nfields = (uint16_t)data;
+    unsigned short nfields = (unsigned short)data;
     ngx_log_debug1(NGX_LOG_DEBUG_HTTP, s->connection->log, 0, "%i", nfields);
     return NGX_OK;
 }
@@ -211,7 +211,7 @@ static ngx_int_t ngx_pg_parser_parse(ngx_pg_save_t *s) {
 }
 
 static ngx_int_t ngx_pg_parser_pid(ngx_pg_save_t *s, const uintptr_t data) {
-    uint32_t pid = (uint32_t)data;
+    unsigned long pid = (unsigned long)data;
     ngx_log_debug1(NGX_LOG_DEBUG_HTTP, s->connection->log, 0, "%i", pid);
     return NGX_OK;
 }
@@ -237,7 +237,7 @@ static ngx_int_t ngx_pg_parser_status_key(ngx_pg_save_t *s, size_t len, const u_
 }
 
 static ngx_int_t ngx_pg_parser_status(ngx_pg_save_t *s, const uintptr_t data) {
-    uint32_t length = (uint32_t)data;
+    unsigned long length = (unsigned long)data;
     ngx_log_debug1(NGX_LOG_DEBUG_HTTP, s->connection->log, 0, "%i", length);
     if (!length) { ngx_log_error(NGX_LOG_ERR, s->connection->log, 0, "!length"); return NGX_ERROR; }
 //    ngx_pg_status_t *status;
@@ -260,25 +260,25 @@ static ngx_int_t ngx_pg_parser_status_val(ngx_pg_save_t *s, size_t len, const u_
 }
 
 static ngx_int_t ngx_pg_parser_tableid(ngx_pg_save_t *s, const uintptr_t data) {
-    uint32_t tableid = (uint32_t)data;
+    unsigned long tableid = (unsigned long)data;
     ngx_log_debug1(NGX_LOG_DEBUG_HTTP, s->connection->log, 0, "%i", tableid);
     return NGX_OK;
 }
 
 static ngx_int_t ngx_pg_parser_ntups(ngx_pg_save_t *s, const uintptr_t data) {
-    uint16_t ntups = (uint16_t)data;
+    unsigned short ntups = (unsigned short)data;
     ngx_log_debug1(NGX_LOG_DEBUG_HTTP, s->connection->log, 0, "%i", ntups);
     return NGX_OK;
 }
 
 static ngx_int_t ngx_pg_parser_typid(ngx_pg_save_t *s, const uintptr_t data) {
-    uint32_t typid = (uint32_t)data;
+    unsigned long typid = (unsigned long)data;
     ngx_log_debug1(NGX_LOG_DEBUG_HTTP, s->connection->log, 0, "%i", typid);
     return NGX_OK;
 }
 
 static ngx_int_t ngx_pg_parser_typlen(ngx_pg_save_t *s, const uintptr_t data) {
-    uint16_t typlen = (uint16_t)data;
+    unsigned short typlen = (unsigned short)data;
     ngx_log_debug1(NGX_LOG_DEBUG_HTTP, s->connection->log, 0, "%i", typlen);
     return NGX_OK;
 }
